language/python/ruamel.yaml/requirements.txt
ruamel.yaml~=0.18.6
pydantic
